added dynamic routes for traiding pairs
