+++ src/app/components/PromoBanner.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import "react";
 import { useState, useEffect } from "react";
