+++ src/app/trade/page.tsx
@@ -1,3 +1,15 @@
-export default function Markets() {
-  return <div className="col-span-full p-4">Trade</div>;
+"use client";
+
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+
+export default function HomePage() {
+  const router = useRouter();
+
+  useEffect(() => {
+    router.replace("/trade/DEXTR/XRD");
+  }, [router]);
+
+  // Render nothing since we are redirecting for now without a home page
+  return;
 }
